Basic integration test tooling/dirs, extract load_toml, extract error_chain usage
