+++ tests/fixtures/test_01.toml
@@ -1,0 +1,2 @@
+[cool]
+yes = true
