+++ tests/load_toml_from_file.rs
@@ -1,0 +1,5 @@
+#[test]
+fn load_toml_from_file() {
+    // TODO: Test something non-trivial
+    assert_eq!(1, 1);
+}
